Refactor file structure and remove redundant files
Refactor file structure and consolidate code into fewer files.

* **index.html**
  - Remove redundant script tags for `chunk-generator.js`, `chunk-manager.js`, `chunk-component.js`, `loading-screen.js`, and `player-controls.js`.
  - Add script tags for `game.js`, `player.js`, `managers.js`, `components.js`, and `generators.js`.
  - Move inline styles to `styles.css`.

* **README.md**
  - Update file structure in the documentation to reflect the new structure.

* **src/game.js**
  - Combine and refactor code from `chunk-generator.js`, `chunk-manager.js`, and `world-manager.js`.
  - Include necessary imports and initialization for the game.

* **src/player.js**
  - Combine and refactor code from `player-controls.js`.
  - Include necessary imports and initialization for player controls.

* **src/managers.js**
  - Combine and refactor code from `chunk-manager.js`, `block-manager.js`, and `voxel-manager.js`.
  - Include necessary imports and initialization for managers.

* **src/components.js**
  - Combine and refactor code from `chunk-component.js`, `loading-screen.js`, and `voxel-component.js`.
  - Include necessary imports and initialization for components.

* **src/generators.js**
  - Combine and refactor code from `texture-generator.js` and `texture-manager.js`.
  - Include necessary imports and initialization for generators.

* **src/styles.css**
  - Move inline styles from `index.html` to `styles.css`.
  - Ensure all styles are properly formatted and organized.

* **Deleted Files**
  - Remove `src/blocks/block-types.js`, `src/components/chunk-component.js`, `src/components/loading-screen.js`, `src/components/player-controls.js`, `src/components/voxel-component.js`, `src/utils/texture-generator.js`, `src/utils/texture-manager.js`, `src/world/chunk-generator.js`, `src/world/chunk-manager.js`, `src/world/voxel-manager.js`, `src/world/world-manager.js`.
